Add dual namespace architecture to SDK (client.store / client.admin)
Restructure @spree/sdk to support both Store and Admin API namespaces,
preparing for upcoming Admin API endpoints. All Store API methods now
live under client.store.* and the Admin API placeholder is at client.admin.*.
