+++ packages/sdk/console.ts
@@ -2,15 +2,15 @@ import * as repl from 'node:repl';
 import { createSpreeClient, SpreeError } from './src';
 
 const baseUrl = process.env.SPREE_URL || 'http://localhost:3000';
-const apiKey = process.env.SPREE_API_KEY || '';
+const publishableKey = process.env.SPREE_PUBLISHABLE_KEY || '';
 
-if (!apiKey) {
-  console.log('Usage: SPREE_API_KEY=spree_pk_xxx npx tsx console.ts');
-  console.log('       SPREE_API_KEY=spree_pk_xxx SPREE_URL=https://api.mystore.com npx tsx console.ts');
+if (!publishableKey) {
+  console.log('Usage: SPREE_PUBLISHABLE_KEY=spree_pk_xxx npx tsx console.ts');
+  console.log('       SPREE_PUBLISHABLE_KEY=spree_pk_xxx SPREE_URL=https://api.mystore.com npx tsx console.ts');
   console.log('');
 }
 
-const client = createSpreeClient({ baseUrl, apiKey });
+const client = createSpreeClient({ baseUrl, publishableKey });
 
 console.log('Spree SDK Console');
 console.log(`Connected to: ${baseUrl}`);
@@ -20,10 +20,10 @@ console.log('  client              - SpreeClient instance');
 console.log('  createSpreeClient   - Create new client');
 console.log('');
 console.log('Examples:');
-console.log('  await client.store.get()');
-console.log('  await client.products.list()');
-console.log('  await client.products.get("my-product", {}, { locale: "fr" })');
-console.log('  await client.taxons.get("categories/clothing", {}, { currency: "EUR" })');
+console.log('  await client.store.store.get()');
+console.log('  await client.store.products.list()');
+console.log('  await client.store.products.get("my-product", {}, { locale: "fr" })');
+console.log('  await client.store.taxons.get("categories/clothing", {}, { currency: "EUR" })');
 console.log('');
 
 const r = repl.start({ prompt: 'spree> ', useGlobal: true });
