+++ packages/sdk/src/index.ts
@@ -1,6 +1,14 @@
 // Main client
-export { SpreeClient, SpreeError, createSpreeClient } from './client';
-export type { SpreeClientConfig, RequestOptions, RetryConfig } from './client';
+export { SpreeClient, createSpreeClient } from './client';
+export type { SpreeClientConfig } from './client';
+
+// Sub-clients
+export { StoreClient } from './store-client';
+export { AdminClient } from './admin-client';
+
+// Request infrastructure
+export { SpreeError } from './request';
+export type { RequestOptions, RetryConfig } from './request';
 
 // All types
 export * from './types';
