+++ packages/sdk/tests/cart.test.ts
@@ -8,28 +8,28 @@ describe('cart', () => {
 
   describe('get', () => {
     it('returns the current cart', async () => {
-      const result = await client.cart.get({ token: 'user-jwt' });
+      const result = await client.store.cart.get({ token: 'user-jwt' });
 
       expect(result.id).toBe('order_1');
       expect(result.token).toBeDefined();
     });
 
     it('works with order token for guest checkout', async () => {
-      const result = await client.cart.get({ orderToken: 'guest-token' });
+      const result = await client.store.cart.get({ orderToken: 'guest-token' });
       expect(result).toBeDefined();
     });
   });
 
   describe('create', () => {
     it('creates a new cart', async () => {
-      const result = await client.cart.create();
+      const result = await client.store.cart.create();
       expect(result.token).toBeDefined();
     });
   });
 
   describe('associate', () => {
     it('associates guest cart with authenticated user', async () => {
-      const result = await client.cart.associate({
+      const result = await client.store.cart.associate({
         token: 'user-jwt',
         orderToken: 'guest-token',
       });
