+++ packages/sdk/tests/orders.test.ts
@@ -9,7 +9,7 @@ describe('orders', () => {
 
   describe('list', () => {
     it('returns paginated orders', async () => {
-      const result = await client.orders.list(undefined, opts);
+      const result = await client.store.orders.list(undefined, opts);
       expect(result.data).toHaveLength(1);
       expect(result.meta.page).toBe(1);
     });
@@ -17,21 +17,21 @@ describe('orders', () => {
 
   describe('create', () => {
     it('creates a new order and returns order_token', async () => {
-      const result = await client.orders.create();
+      const result = await client.store.orders.create();
       expect(result.order_token).toBe('new-order-token');
     });
   });
 
   describe('get', () => {
     it('returns an order by number', async () => {
-      const result = await client.orders.get('R123456', undefined, opts);
+      const result = await client.store.orders.get('R123456', undefined, opts);
       expect(result.number).toBe('R123456');
     });
   });
 
   describe('update', () => {
     it('updates order attributes', async () => {
-      const result = await client.orders.update(
+      const result = await client.store.orders.update(
         'order_1',
         { email: 'new@example.com' },
         opts
@@ -42,24 +42,24 @@ describe('orders', () => {
 
   describe('checkout flow', () => {
     it('advances to next step', async () => {
-      const result = await client.orders.next('order_1', opts);
+      const result = await client.store.orders.next('order_1', opts);
       expect(result.state).toBe('address');
     });
 
     it('advances through all steps', async () => {
-      const result = await client.orders.advance('order_1', opts);
+      const result = await client.store.orders.advance('order_1', opts);
       expect(result.state).toBe('complete');
     });
 
     it('completes the order', async () => {
-      const result = await client.orders.complete('order_1', opts);
+      const result = await client.store.orders.complete('order_1', opts);
       expect(result.state).toBe('complete');
     });
   });
 
   describe('lineItems', () => {
     it('creates a line item', async () => {
-      const result = await client.orders.lineItems.create(
+      const result = await client.store.orders.lineItems.create(
         'order_1',
         { variant_id: 'var_1', quantity: 2 },
         opts
@@ -68,7 +68,7 @@ describe('orders', () => {
     });
 
     it('updates a line item', async () => {
-      const result = await client.orders.lineItems.update(
+      const result = await client.store.orders.lineItems.update(
         'order_1',
         'li_1',
         { quantity: 5 },
@@ -79,14 +79,14 @@ describe('orders', () => {
 
     it('deletes a line item', async () => {
       await expect(
-        client.orders.lineItems.delete('order_1', 'li_1', opts)
+        client.store.orders.lineItems.delete('order_1', 'li_1', opts)
       ).resolves.toBeUndefined();
     });
   });
 
   describe('couponCodes', () => {
     it('applies a coupon code', async () => {
-      const result = await client.orders.couponCodes.apply(
+      const result = await client.store.orders.couponCodes.apply(
         'order_1',
         'SAVE10',
         opts
@@ -95,7 +95,7 @@ describe('orders', () => {
     });
 
     it('removes a coupon code', async () => {
-      const result = await client.orders.couponCodes.remove(
+      const result = await client.store.orders.couponCodes.remove(
         'order_1',
         'promo_1',
         opts
@@ -106,12 +106,12 @@ describe('orders', () => {
 
   describe('shipments', () => {
     it('lists shipments', async () => {
-      const result = await client.orders.shipments.list('order_1', opts);
+      const result = await client.store.orders.shipments.list('order_1', opts);
       expect(result.data).toBeDefined();
     });
 
     it('updates a shipment', async () => {
-      const result = await client.orders.shipments.update(
+      const result = await client.store.orders.shipments.update(
         'order_1',
         'ship_1',
         { selected_shipping_rate_id: 'rate_1' },
@@ -123,12 +123,12 @@ describe('orders', () => {
 
   describe('storeCredits', () => {
     it('adds store credit', async () => {
-      const result = await client.orders.addStoreCredit('order_1', 10, opts);
+      const result = await client.store.orders.addStoreCredit('order_1', 10, opts);
       expect(result).toBeDefined();
     });
 
     it('removes store credit', async () => {
-      const result = await client.orders.removeStoreCredit('order_1', opts);
+      const result = await client.store.orders.removeStoreCredit('order_1', opts);
       expect(result).toBeDefined();
     });
   });
