+++ packages/sdk/tests/products.test.ts
@@ -9,7 +9,7 @@ describe('products', () => {
 
   describe('list', () => {
     it('returns paginated products', async () => {
-      const result = await client.products.list();
+      const result = await client.store.products.list();
 
       expect(result.data).toHaveLength(1);
       expect(result.data[0].name).toBe('Test Product');
@@ -18,7 +18,7 @@ describe('products', () => {
     });
 
     it('passes query parameters', async () => {
-      const result = await client.products.list({
+      const result = await client.store.products.list({
         page: 2,
         per_page: 10,
         'q[name_cont]': 'shirt',
@@ -32,21 +32,21 @@ describe('products', () => {
 
   describe('get', () => {
     it('returns a product by slug', async () => {
-      const result = await client.products.get('test-product');
+      const result = await client.store.products.get('test-product');
 
       expect(result.name).toBe(fixtures.product.name);
       expect(result.slug).toBe('test-product');
     });
 
     it('returns a product by ID', async () => {
-      const result = await client.products.get('prod_1');
+      const result = await client.store.products.get('prod_1');
       expect(result.id).toBe('prod_1');
     });
   });
 
   describe('filters', () => {
     it('returns filter options', async () => {
-      const result = await client.products.filters();
+      const result = await client.store.products.filters();
 
       expect(result.filters).toBeDefined();
       expect(result.sort_options).toBeDefined();
